Move BodyWithBorrowckFacts to consumers
